funcionando

+++ trabalho-pt2-nao_consigo_dar-push (1)/trabalho-pt2-nao_consigo_dar-push/controladores/controlador_cargo.py
@@ -48,14 +48,22 @@ class ControladorCargo:
 
     def excluir_cargo(self):
         cargo = self.seleciona_cargo()
+        if cargo is None:
+            return
         self.__cargo_dao.remove(cargo.id)
-        self.listar_cargos_comuns()
         self.__tela_cargo.exclui_id(cargo.id)
 
     def seleciona_cargo(self):
-        id = self.__tela_cargo.pega_cargo(self.__cargo_dao.get_all())
-        cargo = self.__cargo_dao.get(id)
+        cargo_id = self.__tela_cargo.pega_cargo(self.__cargo_dao.get_all())
+        if cargo_id is None:
+            return cargo_id
+        cargo = self.pega_cargo_por_id(cargo_id)
         return cargo
+
+    def pega_cargo_por_id(self, id: int):
+        for cargo in self.__cargo_dao.get_all():
+            if cargo.id == id:
+                return cargo
 
     def retornar(self):
         self.__control_sistema.inicializa_sistema()
